add meter library
